Fix test coverage
Traced the one statement that was not covered to a bug in pytest and/or
python: The continue statement is correctly executed but not flagged as
executed. When setting a breakpoint on the statement it is not reached.
With the added assignment statement the test coverage report is now ok.