Fix options
Get rid of --radial-distance option: This originated from a
misunderstanding what the original Basic code call 'radial distance':
This is used in the context of the absolute far field in V/m and
specifies the distance of measurment points.
Fix ordering of -T --timing option.
Clarify help-text for --ff-distance and --ff-power options.