Implement Arc geometry object
We now compute the segments *before* the ground. This is needed because
the minimum segment length determines segment end matching and ground
matching. Add compute_segments method to Geo_Container.
Move many methods from Wire to base class Geobj.
In compute_connections fix the case that an object is connected to
itself (a 360° Arc object!)
The compute_ground method is now in Geobj but the derived classes have
their own checks.
When printing print the Geobj name not hard-coded 'WIRE'.
Fix __str__ method of Connected_Ojb, this would sometimes return an
empty string.