Make geo-transformations roundtrip
Now the command-line parameters specifying geo transformations
(rotation, translation, scaling) are reproduced when printing
command-line options. The original un-transformed geometry objects are
printed along with the original geo transformations.

This is necessary for reproducing meaningful helix and arc geometry
objects: This almost always make sense only with a proper geo
transformation but the original geo-object can not specify that
transformation.